FIX: auto cali send failed
and cannot set cali step to 0.001
jira: STUDIO-14042 & STUDIO-14119

(cherry picked from commit c16925931c642d8e27d739a401148bc94c5f3965)

+++ src/slic3r/Utils/CalibUtils.hpp
@@ -10,8 +10,8 @@ class ProgressIndicator;
 class Preset;
 
 namespace GUI {
-extern const float MIN_PA_K_VALUE;
-extern const float MAX_PA_K_VALUE;
+extern const double MIN_PA_K_VALUE;
+extern const double MAX_PA_K_VALUE;
 
 class CalibInfo
 {
